+++ crates/ruma-common/src/identifiers/room_or_alias_id.rs
@@ -2,7 +2,7 @@
 
 use std::hint::unreachable_unchecked;
 
-use ruma_macros::IdDst;
+use ruma_macros::ruma_id;
 
 use super::{RoomAliasId, RoomId, server_name::ServerName};
 
@@ -13,25 +13,21 @@ use super::{RoomAliasId, RoomId, server_name::ServerName};
 /// string slice, the variant is determined by the leading sigil character.
 ///
 /// ```
-/// # use ruma_common::RoomOrAliasId;
-/// assert_eq!(<&RoomOrAliasId>::try_from("#ruma:example.com").unwrap(), "#ruma:example.com");
+/// use ruma_common::RoomOrAliasId;
 ///
-/// assert_eq!(
-///     <&RoomOrAliasId>::try_from("!n8f893n9:example.com").unwrap(),
-///     "!n8f893n9:example.com"
-/// );
+/// assert_eq!(RoomOrAliasId::try_from("#ruma:example.com").unwrap(), "#ruma:example.com");
+///
+/// assert_eq!(RoomOrAliasId::try_from("!n8f893n9:example.com").unwrap(), "!n8f893n9:example.com");
 /// ```
 ///
 /// It can be converted to a `RoomId` or a `RoomAliasId` using `::try_from()` / `.try_into()`.
-/// For example, `<&RoomId>::try_from(room_or_alias_id)` returns either `Ok(room_id)` or
+/// For example, `RoomId::try_from(room_or_alias_id)` returns either `Ok(room_id)` or
 /// `Err(room_alias_id)`.
 ///
 /// [room ID]: https://spec.matrix.org/v1.19/appendices/#room-ids
 /// [room alias ID]: https://spec.matrix.org/v1.19/appendices/#room-aliases
-#[repr(transparent)]
-#[derive(PartialEq, Eq, PartialOrd, Ord, Hash, IdDst)]
 #[ruma_id(validate = ruma_identifiers_validation::room_id_or_alias_id::validate, smallvec_inline_bytes = 48)]
-pub struct RoomOrAliasId(str);
+pub struct RoomOrAliasId;
 
 impl RoomOrAliasId {
     /// Returns the server name of the room (alias) ID.
@@ -66,34 +62,22 @@ enum Variant {
     RoomAliasId,
 }
 
-impl<'a> From<&'a RoomId> for &'a RoomOrAliasId {
-    fn from(room_id: &'a RoomId) -> Self {
-        RoomOrAliasId::from_borrowed_unchecked(room_id.as_str())
-    }
-}
-
-impl<'a> From<&'a RoomAliasId> for &'a RoomOrAliasId {
-    fn from(room_alias_id: &'a RoomAliasId) -> Self {
-        RoomOrAliasId::from_borrowed_unchecked(room_alias_id.as_str())
-    }
-}
-
-impl From<RoomId> for OwnedRoomOrAliasId {
+impl From<RoomId> for RoomOrAliasId {
     fn from(room_id: RoomId) -> Self {
         unsafe { Self::from_inner_unchecked(room_id.into_inner()) }
     }
 }
 
-impl From<RoomAliasId> for OwnedRoomOrAliasId {
+impl From<RoomAliasId> for RoomOrAliasId {
     fn from(room_alias_id: RoomAliasId) -> Self {
         unsafe { Self::from_inner_unchecked(room_alias_id.into_inner()) }
     }
 }
 
-impl TryFrom<OwnedRoomOrAliasId> for RoomId {
+impl TryFrom<RoomOrAliasId> for RoomId {
     type Error = RoomAliasId;
 
-    fn try_from(id: OwnedRoomOrAliasId) -> Result<RoomId, RoomAliasId> {
+    fn try_from(id: RoomOrAliasId) -> Result<RoomId, RoomAliasId> {
         let variant = id.variant();
         let inner = id.into_inner();
 
@@ -106,10 +90,10 @@ impl TryFrom<OwnedRoomOrAliasId> for RoomId {
     }
 }
 
-impl TryFrom<OwnedRoomOrAliasId> for RoomAliasId {
+impl TryFrom<RoomOrAliasId> for RoomAliasId {
     type Error = RoomId;
 
-    fn try_from(id: OwnedRoomOrAliasId) -> Result<RoomAliasId, RoomId> {
+    fn try_from(id: RoomOrAliasId) -> Result<RoomAliasId, RoomId> {
         let variant = id.variant();
         let inner = id.into_inner();
 
@@ -124,13 +108,13 @@ impl TryFrom<OwnedRoomOrAliasId> for RoomAliasId {
 
 #[cfg(test)]
 mod tests {
-    use super::{OwnedRoomOrAliasId, RoomOrAliasId};
+    use super::RoomOrAliasId;
     use crate::IdParseError;
 
     #[test]
     fn valid_room_id_or_alias_id_with_a_room_alias_id() {
         assert_eq!(
-            <&RoomOrAliasId>::try_from("#ruma:example.com")
+            RoomOrAliasId::try_from("#ruma:example.com")
                 .expect("Failed to create RoomAliasId.")
                 .as_str(),
             "#ruma:example.com"
@@ -140,7 +124,7 @@ mod tests {
     #[test]
     fn valid_room_id_or_alias_id_with_a_room_id() {
         assert_eq!(
-            <&RoomOrAliasId>::try_from("!29fhd83h92h0:example.com")
+            RoomOrAliasId::try_from("!29fhd83h92h0:example.com")
                 .expect("Failed to create RoomId.")
                 .as_str(),
             "!29fhd83h92h0:example.com"
@@ -150,7 +134,7 @@ mod tests {
     #[test]
     fn missing_sigil_for_room_id_or_alias_id() {
         assert_eq!(
-            <&RoomOrAliasId>::try_from("ruma:example.com").unwrap_err(),
+            RoomOrAliasId::try_from("ruma:example.com").unwrap_err(),
             IdParseError::MissingLeadingSigil
         );
     }
@@ -159,7 +143,7 @@ mod tests {
     fn serialize_valid_room_id_or_alias_id_with_a_room_alias_id() {
         assert_eq!(
             serde_json::to_string(
-                <&RoomOrAliasId>::try_from("#ruma:example.com")
+                &RoomOrAliasId::try_from("#ruma:example.com")
                     .expect("Failed to create RoomAliasId.")
             )
             .expect("Failed to convert RoomAliasId to JSON."),
@@ -171,7 +155,7 @@ mod tests {
     fn serialize_valid_room_id_or_alias_id_with_a_room_id() {
         assert_eq!(
             serde_json::to_string(
-                <&RoomOrAliasId>::try_from("!29fhd83h92h0:example.com")
+                &RoomOrAliasId::try_from("!29fhd83h92h0:example.com")
                     .expect("Failed to create RoomId.")
             )
             .expect("Failed to convert RoomId to JSON."),
@@ -182,7 +166,7 @@ mod tests {
     #[test]
     fn deserialize_valid_room_id_or_alias_id_with_a_room_alias_id() {
         assert_eq!(
-            serde_json::from_str::<OwnedRoomOrAliasId>(r##""#ruma:example.com""##)
+            serde_json::from_str::<RoomOrAliasId>(r##""#ruma:example.com""##)
                 .expect("Failed to convert JSON to RoomAliasId"),
             "#ruma:example.com"
         );
@@ -191,7 +175,7 @@ mod tests {
     #[test]
     fn deserialize_valid_room_id_or_alias_id_with_a_room_id() {
         assert_eq!(
-            serde_json::from_str::<OwnedRoomOrAliasId>(r#""!29fhd83h92h0:example.com""#)
+            serde_json::from_str::<RoomOrAliasId>(r#""!29fhd83h92h0:example.com""#)
                 .expect("Failed to convert JSON to RoomId"),
             "!29fhd83h92h0:example.com"
         );